requirements.txt
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.2.0
Werkzeug==0.11.10
XlsxWriter==0.8.4
aniso8601==1.1.0
click==6.6
future==0.15.2
itsdangerous==0.24
lxml==2.2.8
pefile==2016.3.28
pypiwin32==219
python-dateutil==2.5.3
python-pptx==0.5.8
pytz==2016.6.1
six==1.10.0
